👌 IMPROVE: Performance improvment
This commit aims at lowering the amount of redundant calls to charCodeAt
to improve performances.

Ord codes are computed once and stored in an attribute for StateBlock, StateCore and StateInline.
We then check this attribute rather than calling the function.

Transfer ord codes whenever possible between StateCore and StateBlocks
so we don't recompute them.