Minor fix to import commands & save to file
Hello,

I was playing with this package and noticed that there are a couple of warning messages with the current important commands. 

Also one other suggestion is to add some code for people to export the html to a file if they happen to not be working in an interactive python session. Happy to remove this if it doesn't make sense for others.