✨ NEW: Add `inline_definitions` option
This option allows for `definition` token to be inserted into the token stream, at the point where the definition is located in the source text
(in addition to the standard extraction to `env`).
It is not part of the Markdown-It JS,
but is useful for cases where one wishes to capture a "loseless"
syntax tree of the parsed Markdown
(in conjunction with the `store_labels` option).